split single package into multiple in order to separate JS and TS config
