+++ .eslintrc.js
@@ -1,1 +1,4 @@
-module.exports = require("./ts-config.js");
+module.exports = {
+	"extends": ["eslint-config-c74-ts"],
+	root: true
+};
