--- .eslintrc_js.js
@@ -1,1 +1,0 @@
-module.exports = require("./index.js");
