--- index.js
@@ -1,6 +1,0 @@
-// JavaScript Configuration
-// Use by doing:
-// extends: ["c74"]
-const { jsConfig } = require("./lib/index");
-
-module.exports = jsConfig;
