+++ packages/eslint-config-c74-base/package.json
@@ -1,0 +1,28 @@
+{
+	"name": "eslint-config-c74-base",
+	"description": "Base configuration of ESLint Rules according to Cycling '74 guide",
+	"version": "1.0.0",
+	"author": "Cycling '74",
+	"contributors": [
+		{
+			"name": "Florian Demmer",
+			"email": "florian@cycling74.com"
+		}
+	],
+	"license": "MIT",
+	"main": "lib/index.js",
+	"scripts": {
+		"build": "tsc -b",
+		"prepare": "yarn run build"
+	},
+	"devDependencies": {},
+	"dependencies": {},
+	"peerDependencies": {
+		"eslint": "^5.14.1"
+	},
+	"readme": "README.md",
+	"repository": "Cycling74/eslint-config-c74",
+	"files": [
+		"lib/**/*"
+	]
+}
