+++ packages/eslint-config-c74-base/src/index.ts
@@ -1,0 +1,15 @@
+import {
+	envConfig,
+	extendsConfig,
+	parserOptions,
+	plugins,
+	rules
+} from "./config";
+
+export = {
+	"extends": extendsConfig,
+	env: envConfig,
+	parserOptions,
+	plugins,
+	rules
+}
