+++ packages/eslint-config-c74/package.json
@@ -1,0 +1,31 @@
+{
+	"name": "eslint-config-c74",
+	"description": "Packaged configuration of ESLint Rules according to Cycling '74 guide",
+	"version": "1.0.0",
+	"author": "Cycling '74",
+	"contributors": [
+		{
+			"name": "Florian Demmer",
+			"email": "florian@cycling74.com"
+		}
+	],
+	"license": "MIT",
+	"main": "lib/index.js",
+	"scripts": {
+		"build": "tsc -b",
+		"prepare": "yarn run build"
+	},
+	"devDependencies": {},
+	"dependencies": {
+		"babel-eslint": "^10.0.1",
+		"eslint-config-c74-base": "1.0.0"
+	},
+	"peerDependencies": {
+		"eslint": "^5.14.1"
+	},
+	"readme": "README.md",
+	"repository": "Cycling74/eslint-config-c74",
+	"files": [
+		"lib/**/*"
+	]
+}
