+++ packages/eslint-config-c74/tsconfig.json
@@ -1,0 +1,10 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "lib",
+		"rootDir": "src"
+	},
+	"include": [
+		"src/**/*"
+	]
+}
