+++ test/mocha.opts
@@ -2,3 +2,4 @@
 --recursive
 --reporter spec
 --timeout 5000
+test/**/*.ts
