--- ts-config.js
@@ -1,6 +1,0 @@
-// Typescript Configuration
-// extends: ["c74/ts"]
-
-const { tsConfig } = require("./lib/index");
-
-module.exports = tsConfig;
