+++ tsconfig.json
@@ -1,16 +1,11 @@
 {
 	"compilerOptions": {
+		"composite": true,
 		"declaration": true,
 		"esModuleInterop": true,
 		"importHelpers": true,
 		"module": "commonjs",
-		"outDir": "lib",
-		"rootDir": "src",
 		"strict": true,
-		"target": "es2017",
-		"composite": true
-	},
-	"include": [
-		"src/**/*"
-	]
+		"target": "es5"
+	}
 }
